docs: update transfer
