--- docs/demo/transfer/section-batch.jsx
@@ -1,55 +1,0 @@
-import React from 'react'
-import DocViewer from '../../../libs/doc-viewer'
-import Transfer from '../../../components/transfer'
-const prefix = 'transfer-batch'
-const code = `import React from 'react'
-import Transfer from '@hi-ui/hiui/es/transfer'\n
-class Demo extends React.Component {
-  constructor () {
-    super()
-    this.state = {
-      datas: this.randomDatas(),
-      targetKeys: [],
-      disabled: false
-    }
-  }
-  randomDatas () {
-    const arr = []
-    for (let i=1;i<16; i++) {
-      arr.push({
-        id: i,
-        content: '选项'+i,
-        disabled: i%3 === 0
-      })
-    }
-    return arr
-  }
-  onChange (movedKeys) {
-    this.setState({
-      targetKeys: movedKeys,
-      disabled: movedKeys.length > 5
-    })
-  }
-  render () {
-    return (
-      <div>
-        <Transfer
-          type='multiple'
-          title={['批量']}
-          disabled={this.state.disabled}
-          targetIds={this.state.targetKeys}
-          data={this.state.datas}
-          onChange={this.onChange.bind(this)}
-        />
-      </div>
-    )
-  }
-}`
-const DemoBatch = () => (
-  <DocViewer
-    code={code}
-    scope={{ Transfer }}
-    prefix={prefix}
-  />
-)
-export default DemoBatch
